fix feature build
